EF Core A-D queries

+++ benchmarks/DapperEntities/OrderLine.cs
@@ -11,7 +11,7 @@ public class OrderLine
     
     public int PackageTypeID { get; set; }
     
-    public decimal Quantity { get; set; }
+    public int Quantity { get; set; }
     
     public decimal? UnitPrice { get; set; }
     
@@ -21,7 +21,7 @@ public class OrderLine
     
     public DateTime? PickingCompletedWhen { get; set; }
     
-    public decimal LastEditedBy { get; set; }
+    public int LastEditedBy { get; set; }
     
     public DateTime LastEditedWhen { get; set; }
 }
