+++ benchmarks/PetaPocoEntities/OrderLine.cs
@@ -15,7 +15,7 @@ public class OrderLine
 
     public int PackageTypeID { get; set; }
 
-    public decimal Quantity { get; set; }
+    public int Quantity { get; set; }
 
     public decimal? UnitPrice { get; set; }
 
@@ -25,7 +25,7 @@ public class OrderLine
 
     public DateTime? PickingCompletedWhen { get; set; }
 
-    public decimal LastEditedBy { get; set; }
+    public int LastEditedBy { get; set; }
 
     public DateTime LastEditedWhen { get; set; }
 }
